Export storybook middleware
